refactor: improve gio-clipboard files naming and others
