+++ gravitee-apim-console-webui/src/shared/components/gio-clipboard/gio-clipboard-copy-wrapper.component.scss
